AE-2555 Fix Gear Down In Transit and Gear Up In Transit
It was found out that many dataframes trigger spurious Gear (*) Red Warnings.
The end goal is to fix those frames. In the meantime, it's been decided to
filter out the Red Warnings and only use those close to Gear Up or
Gear Down to determine the transition periods.

Also fixed the dtype for those 2 Multistate Parameters to int.
Fixed timing errors for B737 Classic where the number of frames was used
instead of the timing.